requirement.txt
arrow==0.15.5
backpack==0.1
blinker==1.4
Cerberus==1.3.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cleo==0.6.8
click==7.1.1
crypto==1.4.1
cryptography==2.9.2
eloquent==0.5
Faker==0.8.18
Flask==1.1.2
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
idna==2.9
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mongoengine==0.19.1
Naked==0.1.31
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
pyaml==16.12.2
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylev==1.3.0
pymongo==3.10.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
pytzdata==2019.3
PyYAML==5.3.1
requests==2.23.0
shellescape==3.8.1
simplejson==3.17.0
six==1.14.0
text-unidecode==1.2
tzlocal==1.5.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.2.1
